refactor: varient -> variant
