+++ lib/json/experiment_variant.dart
